pyproject.toml
[project]
name = "otp-sync-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=44.0.1",
    "fastapi>=0.115.8",
    "pycryptodome>=3.21.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "telebot>=0.0.5",
    "uvicorn>=0.34.0",
]
